Renamed FileReader to ArchiveReader
Because that is what it is now.
No more weird hybrid of a stream and a c-library.

The one place where the "simple filereading"  was used is replaced by just using a stream directly
